tests/wpt/tests/css/css-grid/parsing/grid-row-shortest-serialization.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="author" title="Sammy Gill" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-2/#placement-shorthands"/>
<meta name="assert" content="grid-row computed value should be the shortest serialization possible"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
    test_computed_value("grid-row", "auto / auto", "auto");
    test_computed_value("grid-row", "auto", "auto");
    test_computed_value("grid-row", "10 / auto", "10");
    test_computed_value("grid-row", "10");
    test_computed_value("grid-row", "-10 / auto", "-10");
    test_computed_value("grid-row", "-10");
    test_computed_value("grid-row", "span 2 / auto", "span 2");
    test_computed_value("grid-row", "span 2");
    test_computed_value("grid-row", "3 last / auto", "3 last");
    test_computed_value("grid-row", "3 last");
    test_computed_value("grid-row", "span first / auto", "span first");
    test_computed_value("grid-row", "span first");
    test_computed_value("grid-row", "span 2 first / auto", "span 2 first");
    test_computed_value("grid-row", "span 2 first");
    test_computed_value("grid-row", "last / last", "last");
</script>
</body>
</html>
